requirements.txt
python-docx==0.8.10
pandas==0.23.4